Refactored out config locator into own class
