+++ src/Syringe.Core/Configuration/IConfigLocator.cs
@@ -1,0 +1,7 @@
+﻿namespace Syringe.Core.Configuration
+{
+    public interface IConfigLocator
+    {
+        string ResolveConfigFile(string fileName);
+    }
+}
