+++ src/Syringe.Core/Configuration/IConfigurationStore.cs
@@ -3,7 +3,5 @@ namespace Syringe.Core.Configuration
 	public interface IConfigurationStore
 	{
 	    IConfiguration Load();
-	    string ResolveConfigFile(string fileName);
-
 	}
 }
