+++ src/Syringe.Tests/Syringe.Tests.csproj
@@ -221,6 +221,7 @@
     <Compile Include="Integration\ClientAndService\EnvironmentsClientTests.cs" />
     <Compile Include="Integration\ClientAndService\Helpers.cs" />
     <Compile Include="Integration\ClientAndService\TaskClientTests.cs" />
+    <Compile Include="Integration\Core\Configuration\ConfigLocatorTests.cs" />
     <Compile Include="Integration\Core\Configuration\JsonConfigurationStoreTests.cs" />
     <Compile Include="Integration\Core\Environments\JsonEnvironmentProviderTests.cs" />
     <Compile Include="Integration\Core\IO\FileHandlerTests.cs" />
